refactor: service에서 dto 반환하도록 수정
- 도메인을 노출할 경우 결합도가 높아진다.
