+++ src/main/java/roomescape/member/dto/request/SignupRequest.java
@@ -1,4 +1,4 @@
-package roomescape.member.dto;
+package roomescape.member.dto.request;
 
 public record SignupRequest(String email, String password, String name) {
 }
