+++ src/main/java/roomescape/member/dto/response/SignUpResponse.java
@@ -1,0 +1,10 @@
+package roomescape.member.dto.response;
+
+import roomescape.member.domain.Member;
+
+public record SignUpResponse(Long id) {
+
+    public static SignUpResponse from(final Member member) {
+        return new SignUpResponse(member.getId());
+    }
+}
