+++ src/main/java/roomescape/member/service/MemberService.java
@@ -6,8 +6,9 @@ import roomescape.global.auth.dto.UserInfo;
 import roomescape.global.auth.service.MyPasswordEncoder;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRole;
-import roomescape.member.dto.MemberResponse;
-import roomescape.member.dto.SignupRequest;
+import roomescape.member.dto.request.SignupRequest;
+import roomescape.member.dto.response.MemberResponse;
+import roomescape.member.dto.response.SignUpResponse;
 import roomescape.member.exception.MemberDuplicatedException;
 import roomescape.member.exception.MemberNotFoundException;
 import roomescape.member.repository.MemberRepository;
@@ -23,13 +24,13 @@ public class MemberService {
         this.myPasswordEncoder = myPasswordEncoder;
     }
 
-    public Member signup(final SignupRequest signupRequest) {
+    public SignUpResponse signup(final SignupRequest signupRequest) {
         String encodedPassword = myPasswordEncoder.encode(signupRequest.password());
         Member member = new Member(signupRequest.name(), signupRequest.email(), encodedPassword, MemberRole.USER);
         if (memberRepository.existsByEmail(signupRequest.email())) {
             throw new MemberDuplicatedException("이미 존재하는 회원입니다.");
         }
-        return memberRepository.save(member);
+        return SignUpResponse.from(memberRepository.save(member));
     }
 
     public List<MemberResponse> findAllUsers() {
@@ -40,6 +41,6 @@ public class MemberService {
 
     public Member getMember(final UserInfo userInfo) {
         return memberRepository.findById(userInfo.id())
-                .orElseThrow(()->new MemberNotFoundException("존재하지 않은 멤버입니다."));
+                .orElseThrow(() -> new MemberNotFoundException("존재하지 않은 멤버입니다."));
     }
 }
