+++ src/test/java/roomescape/MissionStepTest.java
@@ -23,8 +23,8 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.TestPropertySource;
 import roomescape.global.auth.dto.CheckLoginResponse;
 import roomescape.global.auth.dto.LoginRequest;
-import roomescape.member.dto.MemberResponse;
-import roomescape.member.dto.SignupRequest;
+import roomescape.member.dto.response.MemberResponse;
+import roomescape.member.dto.request.SignupRequest;
 import roomescape.reservation.controller.ReservationController;
 import roomescape.reservation.domain.ReservationStatus;
 import roomescape.reservation.dto.response.MyReservationResponse;
