+++ src/test/java/roomescape/member/service/MemberServiceTest.java
@@ -12,8 +12,9 @@ import org.springframework.context.annotation.Import;
 import org.springframework.test.context.TestPropertySource;
 import roomescape.global.auth.service.MyPasswordEncoder;
 import roomescape.member.domain.Member;
-import roomescape.member.dto.MemberResponse;
-import roomescape.member.dto.SignupRequest;
+import roomescape.member.dto.request.SignupRequest;
+import roomescape.member.dto.response.MemberResponse;
+import roomescape.member.dto.response.SignUpResponse;
 import roomescape.member.repository.MemberRepository;
 
 @DataJpaTest
@@ -38,18 +39,18 @@ public class MemberServiceTest {
 
     @Test
     void signUpTest() {
-        Member member = memberService.signup(new SignupRequest("userr@gmail.com", "password", "userr"));
+        SignUpResponse response = memberService.signup(new SignupRequest("userr@gmail.com", "password", "userr"));
 
-        Optional<Member> optionalMember = memberRepository.findById(member.getId());
+        Optional<Member> optionalMember = memberRepository.findById(response.id());
         assertThat(optionalMember.get().getName()).isEqualTo("userr");
     }
 
     @Test
     void findAllUsersTest() {
-        Member member = memberService.signup(new SignupRequest("userr@gmail.com", "password", "userr"));
-        Member member2 = memberService.signup(new SignupRequest("userrr@gmail.com", "password", "userrr"));
-        Member member3 = memberService.signup(new SignupRequest("userrrr@gmail.com", "password", "userrrr"));
-        List<MemberResponse> memberResponses= memberService.findAllUsers();
+        memberService.signup(new SignupRequest("user1@gmail.com", "password", "user1"));
+        memberService.signup(new SignupRequest("user2@gmail.com", "password", "user2"));
+        memberService.signup(new SignupRequest("user3@gmail.com", "password", "user3"));
+        List<MemberResponse> memberResponses = memberService.findAllUsers();
         assertThat(memberResponses.size()).isEqualTo(3);
     }
 }
